requirements.txt
numpy==1.21.5
opencv-python-headless
Pillow==9.4.0
spacy
torchvision
jsonlines==3.1.0